pyproject.toml
[tool.poetry]
name = "brew-dependencies"
version = "2024"
description = "Dependencies for development"
package-mode = false
authors = ["Brew Foundation"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
ttkbootstrap = "^1.10.1"
supabase = "^2.4.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
